Return decoded byte size

+++ include/b64.h
@@ -20,9 +20,9 @@ size_t b64_encode(char* encoded_str, const uint8_t* input_bytes, const size_t in
  * @param[out] decoded_bytes Base64-decoded byte array
  * @param[in] input_str Input string, NULL-terminated
  * @param[in] input_size_in_bytes Input size in bytes
- * @retval 0 Decoding succeeded
- * @retval 1 Decoding failed
+ * @return Byte size of the encoded string
+ * @retval 0 Decoding failed
  */
-int b64_decode(uint8_t* decoded_bytes, const char* input_str, const size_t input_size_in_bytes);
+size_t b64_decode(uint8_t* decoded_bytes, const char* input_str, const size_t input_size_in_bytes);
 
 #endif // B64_H
